+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/mechanisms/Transfer.java
@@ -2,6 +2,8 @@ package org.firstinspires.ftc.teamcode.mechanisms;
 
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
+import org.firstinspires.ftc.teamcode.mechanisms.tests.QQ_Test;
+
 import java.util.List;
 
 class Transfer implements QQ_Mechanism {
