+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/mechanisms/tests/QQ_Test.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.mechanisms;
+package org.firstinspires.ftc.teamcode.mechanisms.tests;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
