+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/mechanisms/tests/QQ_TestAnalogSensor.java
@@ -1,10 +1,8 @@
-package org.firstinspires.ftc.teamcode.mechanisms;
+package org.firstinspires.ftc.teamcode.mechanisms.tests;
 
 import com.qualcomm.robotcore.hardware.AnalogInput;
-import com.qualcomm.robotcore.hardware.DistanceSensor;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.robotcore.external.navigation.DistanceUnit;
 
 public class QQ_TestAnalogSensor extends QQ_Test{
     private AnalogInput analogInput;
