+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/mechanisms/tests/QQ_TestMotor.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.mechanisms;
+package org.firstinspires.ftc.teamcode.mechanisms.tests;
 
 import com.qualcomm.robotcore.hardware.DcMotor;
