+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/mechanisms/tests/QQ_TestServo.java
@@ -1,6 +1,5 @@
-package org.firstinspires.ftc.teamcode.mechanisms;
+package org.firstinspires.ftc.teamcode.mechanisms.tests;
 
-import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Servo;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
